revert login page refactor
