+++ public/locales/en/translation.json
@@ -4,7 +4,6 @@
   "Email": "Email",
   "you@example.com": "you@example.com",
   "Password": "Password",
-  "••••••••": "••••••••",
   "Sign Up": "Sign Up",
   "Forgot your password?": "Forgot your password?",
   "Reset": "Reset",
