+++ public/locales/es/translation.json
@@ -4,7 +4,6 @@
   "Email": "Correo electrónico",
   "you@example.com": "tu@ejemplo.com",
   "Password": "Contraseña",
-  "••••••••": "••••••••",
   "Sign Up": "Registrarse",
   "Forgot your password?": "¿Olvidaste tu contraseña?",
   "Reset": "Restablecer",
